requirements.txt
python-dotenv>=1.0.0
requests>=2.31.0
flask>=2.3.0
pyyaml>=6.0
openai>=1.0.0
anthropic>=0.7.0
flask-cors>=4.0.0